RGW/logging: add filtering for bucket logging
