+++ doc/radosgw/s3/bucketops.rst
@@ -751,6 +751,26 @@ Parameters are XML encoded in the body of the request, in the following format:
       <TargetPrefix>string</TargetPrefix>
       <LoggingType>Standard|Journal</LoggingType>
       <ObjectRollTime>integer</ObjectRollTime>
+      <Filter>
+        <S3Key>
+          <FilterRule>
+            <Name>suffix/prefix/regex</Name>
+            <Value></Value>
+          </FilterRule>
+        </S3Key>
+        <S3Metadata>
+          <FilterRule>
+            <Name></Name>
+            <Value></Value>
+          </FilterRule>
+        </S3Metadata>
+        <S3Tags>
+          <FilterRule>
+            <Name></Name>
+            <Value></Value>
+          </FilterRule>
+        </S3Tags>
+      </Filter>
     </LoggingEnabled>
   </BucketLoggingStatus>
 
@@ -881,6 +901,26 @@ Response is XML encoded in the body of the request, in the following format:
       <TargetPrefix>string</TargetPrefix>
       <LoggingType>Standard|Journal</LoggingType>
       <ObjectRollTime>integer</ObjectRollTime>
+      <Filter>
+        <S3Key>
+          <FilterRule>
+            <Name>suffix/prefix/regex</Name>
+            <Value></Value>
+          </FilterRule>
+        </S3Key>
+        <S3Metadata>
+          <FilterRule>
+            <Name></Name>
+            <Value></Value>
+          </FilterRule>
+        </S3Metadata>
+        <S3Tags>
+          <FilterRule>
+            <Name></Name>
+            <Value></Value>
+          </FilterRule>
+        </S3Tags>
+      </Filter>
     </LoggingEnabled>
   </BucketLoggingStatus>
