+++ examples/rgw/boto3/service-2.sdk-extras.json
@@ -310,6 +310,10 @@
                 "RecordsBatchSize":{
                     "shape":"RecordsBatchSize",
                     "documentation":"indicates how many records to batch in memory before writing to the object. if set to zero, records are written syncronously to the object. if <code>ObjectRollTime</code>e is reached, the batch of records will be written to the object regardless of the number of records. </p>"
+                },
+                "Filter":{
+                    "shape":"NotificationConfigurationFilter",
+                    "documentation":"<p>A filter for all log object. Types of filter for the object by its: attributes, tags and key (prefix, suffix and regex).</p>"
                 }
             },
             "documentation":"<p>Describes where logs are stored the prefix assigned to all log object keys for a bucket, and their format. also, the level the delivery guarantee of the records.</p>"
